Use uint32 for freq. values to support 2.4 GHz freqs on 32 bit.
Previously the int type was used which would result in compile errors
on 32bit platforms since LoRaWAN 2.4 GHz frequency values do not fit
within int32 type (signed).

+++ cmd/chirpstack-network-server/cmd/root_run.go
@@ -130,7 +130,7 @@ func setRXParameters() error {
 	}
 
 	if config.C.NetworkServer.NetworkSettings.RX2Frequency == -1 {
-		config.C.NetworkServer.NetworkSettings.RX2Frequency = defaults.RX2Frequency
+		config.C.NetworkServer.NetworkSettings.RX2Frequency = int64(defaults.RX2Frequency)
 	}
 
 	return nil
